Add finishing touches to Intake and update the example plugin
nuff said. I'll be updating the readme pretty soon and might add
the ability to write custom help messages because the /help command
needs some attention. However, I'm pretty occupied with other stuff
right now.
