+++ bukkit/src/main/java/app/ashcon/intake/bukkit/graph/BasicBukkitCommandGraph.java
@@ -5,14 +5,13 @@ import app.ashcon.intake.bukkit.BukkitAuthorizer;
 import app.ashcon.intake.bukkit.BukkitModule;
 import app.ashcon.intake.dispatcher.SimpleDispatcher;
 import app.ashcon.intake.fluent.CommandGraph;
-import app.ashcon.intake.fluent.GroupDispatcherNode;
+import app.ashcon.intake.fluent.DispatcherNode;
 import app.ashcon.intake.parametric.Module;
 import app.ashcon.intake.parametric.ParametricBuilder;
 import app.ashcon.intake.parametric.provider.PrimitivesModule;
 import java.util.Arrays;
-import java.util.function.Function;
 
-public class BasicBukkitCommandGraph extends CommandGraph<GroupDispatcherNode> {
+public class BasicBukkitCommandGraph extends CommandGraph<DispatcherNode> {
 
     /**
      * Create a new command graph with a simple dispatcher node.
@@ -28,6 +27,6 @@ public class BasicBukkitCommandGraph extends CommandGraph<GroupDispatcherNode> {
         Arrays.stream(modules).forEach(builder.getInjector()::install);
 
         setBuilder(builder);
-        setRootDispatcherNode(new GroupDispatcherNode(this, null, new SimpleDispatcher()));
+        setRootDispatcherNode(new DispatcherNode(this, null, new SimpleDispatcher()));
     }
 }
