+++ core/src/main/java/app/ashcon/intake/fluent/AbstractDispatcherNode.java
@@ -18,10 +18,7 @@
  */
 package app.ashcon.intake.fluent;
 
-import app.ashcon.intake.CommandCallable;
 import app.ashcon.intake.dispatcher.Dispatcher;
-import app.ashcon.intake.dispatcher.SimpleDispatcher;
-import app.ashcon.intake.parametric.ParametricBuilder;
 
 public abstract class AbstractDispatcherNode {
 
@@ -33,7 +30,7 @@ public abstract class AbstractDispatcherNode {
      * Create a new instance.
      *
      * @param graph      the root fluent graph object
-     * @param parent     the parent node, or null
+     * @param parent     the getParent node, or null
      * @param dispatcher the dispatcher for this node
      */
     protected AbstractDispatcherNode(CommandGraph graph, AbstractDispatcherNode parent, Dispatcher dispatcher) {
@@ -43,64 +40,28 @@ public abstract class AbstractDispatcherNode {
     }
 
     /**
-     * Register a command with this dispatcher.
-     *
-     * @param callable the executor
-     * @param aliases  the list of aliases, where the first alias is the primary one
-     */
-    public void register(CommandCallable callable, String... aliases) {
-        dispatcher.registerCommand(callable, aliases);
-    }
-
-    /**
-     * Build and register a command with this dispatcher using the
-     * {@link ParametricBuilder} assigned on the root {@link CommandGraph}.
-     *
-     * @param object the object provided to the {@link ParametricBuilder}
-     * @return this object
-     * @see ParametricBuilder#registerMethodsAsCommands(Dispatcher, Object)
-     */
-    public AbstractDispatcherNode registerMethods(Object object) {
-        graph.getBuilder().registerMethodsAsCommands(getDispatcher(), object);
-        return this;
-    }
-
-    /**
      * Create a new command that will contain sub-commands.
      *
      * <p>The object returned by this method can be used to add sub-commands. To
-     * return to this "parent" context, use {@link DispatcherNode#graph()}.</p>
+     * return to this "getParent" context, use {@link DispatcherNode#graph}.</p>
      *
      * @param aliases the list of aliases, where the first alias is the primary one
      * @return an object to place sub-commands
      */
-    public DispatcherNode group(String... aliases) {
-        SimpleDispatcher command = new SimpleDispatcher();
-        getDispatcher().registerCommand(command, aliases);
-        return new DispatcherNode(graph, this, command);
-    }
+    public abstract AbstractDispatcherNode registerNode(String... aliases);
 
     /**
-     * Return the parent node.
+     * Return the getParent node.
      *
-     * @return the parent node
-     * @throws RuntimeException if there is no parent node.
+     * @return the getParent node
+     * @throws RuntimeException if there is no getParent node.
      */
-    public AbstractDispatcherNode parent() {
+    public AbstractDispatcherNode getParent() {
         if (parent != null) {
             return parent;
         }
 
-        throw new RuntimeException("This node does not have a parent");
-    }
-
-    /**
-     * Get the root command graph.
-     *
-     * @return the root command graph
-     */
-    public CommandGraph graph() {
-        return graph;
+        throw new RuntimeException("This node does not have a getParent");
     }
 
     /**
