+++ core/src/main/java/app/ashcon/intake/fluent/DispatcherNode.java
@@ -20,6 +20,8 @@ package app.ashcon.intake.fluent;
 
 import app.ashcon.intake.dispatcher.Dispatcher;
 import app.ashcon.intake.dispatcher.SimpleDispatcher;
+import app.ashcon.intake.parametric.ParametricBuilder;
+import java.lang.reflect.Method;
 
 /**
  * A collection of commands.
@@ -30,40 +32,48 @@ public class DispatcherNode extends AbstractDispatcherNode {
      * Create a new instance.
      *
      * @param graph      the root fluent graph object
-     * @param parent     the parent node, or null
+     * @param parent     the getParent node, or null
      * @param dispatcher the dispatcher for this node
      */
-    DispatcherNode(CommandGraph graph, AbstractDispatcherNode parent, Dispatcher dispatcher) {
+    public DispatcherNode(CommandGraph graph, AbstractDispatcherNode parent, Dispatcher dispatcher) {
         super(graph, parent, dispatcher);
     }
 
     /**
-     * {@inheritDoc}
+     * Build and register all commands within a class with this dispatcher using
+     * the {@link ParametricBuilder} assigned on the root {@link CommandGraph}.
+     *
+     * @param object the object who's methods are provided to the {@link ParametricBuilder}
+     * @return this object
+     * @see ParametricBuilder#registerMethodAsCommand(Dispatcher, Method)
      */
-    public DispatcherNode registerMethods(Object object) {
-        return (DispatcherNode) super.registerMethods(object);
+    public AbstractDispatcherNode registerClassMethods(Object object) {
+        for(Method method : object.getClass().getDeclaredMethods())
+            graph.getBuilder().registerMethodAsCommand(getDispatcher(), object, method);
+        return this;
     }
 
     /**
      * Create a new command that will contain sub-commands.
      *
      * <p>The object returned by this method can be used to add sub-commands. To
-     * return to this "parent" context, use {@link DispatcherNode#graph()}.</p>
+     * return to this "getParent" context, use {@link DispatcherNode#graph}.</p>
      *
-     * @param alias the list of aliases, where the first alias is the primary one
+     * @param aliases the list of aliases, where the first aliases is the primary one
      * @return an object to place sub-commands
      */
-    public DispatcherNode group(String... alias) {
+    @Override
+    public DispatcherNode registerNode(String... aliases) {
         SimpleDispatcher command = new SimpleDispatcher();
-        getDispatcher().registerCommand(command, alias);
+        getDispatcher().registerCommand(command, aliases);
         return new DispatcherNode(graph, this, command);
     }
 
     /**
      * {@inheritDoc}
      */
-    public DispatcherNode parent() {
-        return (DispatcherNode) super.parent();
+    public DispatcherNode getParent() {
+        return (DispatcherNode) super.getParent();
     }
 
 }
