--- core/src/main/java/app/ashcon/intake/group/Group.java
@@ -1,40 +1,0 @@
-/*
- * Intake, a command processing library
- * Copyright (C) sk89q <http://www.sk89q.com>
- * Copyright (C) Intake team and contributors
- *
- * This program is free software: you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as published by the
- * Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License
- * for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-package app.ashcon.intake.group;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotates a {@link ElementType#METHOD} or {@link ElementType#TYPE} that should be
- * registered under groups.
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.TYPE})
-public @interface Group {
-
-    /**
-     * The groups to register under.
-     *
-     * @return the groups
-     */
-    At[] value();
-}
