+++ examples/bukkit/src/main/java/app/ashcon/intake/example/ExamplePlugin.java
@@ -1,13 +1,26 @@
 package app.ashcon.intake.example;
 
 import app.ashcon.intake.bukkit.BukkitIntake;
+import app.ashcon.intake.bukkit.graph.BasicBukkitCommandGraph;
+import app.ashcon.intake.fluent.DispatcherNode;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public class ExamplePlugin extends JavaPlugin {
 
     @Override
     public void onLoad() {
-        new BukkitIntake(this, new ExampleCommands());
+        super.onLoad();
+
+        BasicBukkitCommandGraph cmdGraph = new BasicBukkitCommandGraph();
+
+        DispatcherNode testNode = cmdGraph.getRootDispatcherNode().registerNode("test");
+
+        testNode.registerClassMethods(new TestCommands());
+        testNode.registerNode("math").registerClassMethods(new MathCommands());
+
+        BukkitIntake intake = new BukkitIntake(this, cmdGraph);
+
+        intake.register();
     }
 
 }
