--- settings.xml
@@ -1,17 +1,0 @@
-<settings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://maven.apache.org/SETTINGS/1.0.0"
-  xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0
-                      https://maven.apache.org/xsd/settings-1.0.0.xsd">
-  <servers>
-    <server>
-      <id>ashcon-releases</id>
-      <username>${env.REPO_USERNAME}</username>
-      <password>${env.REPO_PASSWORD}</password>
-    </server>
-    <server>
-      <id>ashcon-snapshots</id>
-      <username>${env.REPO_USERNAME}</username>
-      <password>${env.REPO_PASSWORD}</password>
-    </server>
-  </servers>
-</settings>
